requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-multipart>=0.0.9
websockets>=12.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
reportlab>=4.1.0
gymnasium>=0.29.1
openenv-core==0.2.3
trl>=0.8.6
transformers>=4.40.0
datasets>=2.19.0
torch>=2.2.0
# unsloth  # Optional: install separately in GPU training environments. Not required for serving.
accelerate>=0.30.0
openai>=1.25.0
requests>=2.31.0
google-generativeai>=0.5.0
anthropic>=0.25.0
